Fix memory pressure calculation units in Haiku platform code.
This patch corrects the calculation of available memory in `MemoryPressureHandlerHaiku.cpp` and `WebProcessPoolHaiku.cpp`.

Previously, `system_info.free_memory` (which is in bytes) was being treated as pages and/or incorrectly combined with `system_info.cached_pages` (which is in pages).
- In `MemoryPressureHandlerHaiku.cpp`, `free_memory` was added to `cached_pages` before multiplying by `B_PAGE_SIZE`, resulting in a massive overestimation of free memory.
- In `WebProcessPoolHaiku.cpp`, `free_memory` was multiplied by `B_PAGE_SIZE`, also resulting in a massive overestimation.

The fix ensures:
1. `free_memory` is treated as bytes.
2. `cached_pages` is converted to bytes (pages * B_PAGE_SIZE).
3. Both are summed to represent total available memory, consistent with Haiku's aggressive caching behavior where cached memory is readily available.

This change ensures that low-memory warnings and pressure handlers trigger correctly on Haiku.

+++ Source/WebCore/platform/haiku/MemoryPressureHandlerHaiku.cpp
@@ -52,7 +52,7 @@ void MemoryPressureHandler::install()
         system_info info;
         if (get_system_info(&info) == B_OK) {
             // Include cached pages as available memory since Haiku caches aggressively
-            uint64_t freeMemory = (uint64_t)(info.free_memory + info.cached_pages) * B_PAGE_SIZE;
+            uint64_t freeMemory = (uint64_t)info.free_memory + ((uint64_t)info.cached_pages * B_PAGE_SIZE);
             uint64_t totalMemory = (uint64_t)info.max_pages * B_PAGE_SIZE;
 
             // Trigger if less than 64MB or 10% memory free
